// src/features/whatsapp-templates/enums/template.enum.ts
export enum TemplateCategory {
    MARKETING = "MARKETING",
    UTILITY = "UTILITY",
    AUTHENTICATION = "AUTHENTICATION"
}

export enum TemplateStatus {
    APPROVED = "APPROVED",
    PENDING = "PENDING",
    REJECTED = "REJECTED",
    DRAFT = "DRAFT",
    PAUSED = "PAUSED"
}

export enum HeaderType {
    TEXT = 'TEXT',
    IMAGE = 'IMAGE',
    VIDEO = 'VIDEO',
    DOCUMENT = 'DOCUMENT',
}

export enum ButtonType {
    QUICK_REPLY = 'QUICK_REPLY',
    CALL_TO_ACTION = 'CALL_TO_ACTION',
}

export enum ActionType {
    PHONE_NUMBER = 'PHONE_NUMBER',
    URL = 'URL',
}